refactor(editor): extract event dispatch into EventDispatcher handler
Move dispatch_minga_event/4 and all remote/distribution helpers into
MingaEditor.Handlers.EventDispatcher. The Editor GenServer handle_info
clause for {:minga_event, ...} now delegates to EventDispatcher.dispatch/4.

Promotes schedule_render/2, apply_diagnostic_decorations/2, and
update_test_notification/2 from defp to @doc false def so the new
handler can call them.

Editor reduced from 2,609 to 2,095 lines (-514).

Closes #1433 (partial)
